Fixing dependency issue
