+++ Redux/Actions/Comments/CommentsActions.js
@@ -1,23 +1,10 @@
 import {
-    ADD_COMMENT,
-    BULK_COMMENTS_DELETE,
     DELETE_COMMENT,
     FETCH_COMMENTS,
 } from 'Gruvee/Redux/Actions/ActionsType'
-import {
-    AddSongComment,
-    DeleteSongComment,
-} from 'Gruvee/Redux/Actions/Playlists/PlaylistActions'
+import { DeleteSongComment } from 'Gruvee/Redux/Actions/Playlists/SharedPlaylistActions'
 
 // Action Creators
-const addComment = comment => {
-    return {
-        // evjand - "This app will never be finished" (02/07/20)
-        type: ADD_COMMENT,
-        data: comment,
-    }
-}
-
 const deleteComment = commentId => {
     return {
         type: DELETE_COMMENT,
@@ -33,22 +20,6 @@ const fetchComments = comments => {
 }
 
 // Thunks
-export const AddComment = (comment, songId, playlistId) => {
-    // syszen - "it broke na na's, i had hope" (02/09/20)
-    // Make async call to update our db with new comment
-    // We also will need to update our respected song given the songI
-
-    return (dispatch, getState) => {
-        if (comment) {
-            // Storing in PlaylistsDataReducer State
-            dispatch(AddSongComment(comment.id, songId, playlistId))
-
-            // Storing in CommentsDataReducer State
-            dispatch(addComment(comment))
-        }
-    }
-}
-
 export const DeleteComment = (commentId, songId, playlistId) => {
     // Make async call to update our db with removing comment
     // We also will need to update our respected song given the songId
@@ -67,13 +38,5 @@ export const FetchComments = songId => {
     return (dispatch, getState) => {
         const comments = [] // Will be an array of comments objects
         dispatch(fetchComments(comments))
-    }
-}
-
-// Local State Settings
-export const BulkCommentsDelete = commentIds => {
-    return {
-        type: BULK_COMMENTS_DELETE,
-        data: commentIds,
     }
 }
