+++ Redux/Actions/Playlists/PlaylistActions.js
@@ -4,34 +4,17 @@
 // ywknlme - "TODO DODODODODO DODODODODODODODO DODODODODO... *plays Darude – Sandstorm*" (01/29/20)
 import {
     ADD_PLAYLIST,
-    ADD_PLAYLIST_SONG,
-    ADD_SONG_COMMENT,
     DELETE_PLAYLIST,
-    DELETE_PLAYLIST_SONG,
-    DELETE_SONG_COMMENT,
     FETCH_PLAYLISTS,
     SET_CURRENT_PLAYLIST_ID,
-} from 'Gruvee/Redux/Actions/ActionsType'
+} from '../ActionsType'
+import { DeleteSong } from '../Songs/SharedSongActions'
 
 // Action Creators
 const addPlaylist = playlist => {
     return {
         type: ADD_PLAYLIST,
         data: playlist,
-    }
-}
-
-const addPlaylistSong = (songId, playlistId) => {
-    return {
-        type: ADD_PLAYLIST_SONG,
-        data: { songId, playlistId },
-    }
-}
-
-const addSongComment = (commentId, songId, playlistId) => {
-    return {
-        type: ADD_SONG_COMMENT,
-        data: { commentId, songId, playlistId },
     }
 }
 
@@ -42,20 +25,6 @@ const deletePlaylist = (playlistId, playlists) => {
     }
 }
 
-const deletePlaylistSong = (songId, playlistId) => {
-    return {
-        type: DELETE_PLAYLIST_SONG,
-        data: { songId, playlistId },
-    }
-}
-
-const deleteSongComment = (commentId, songId, playlistId) => {
-    return {
-        type: DELETE_SONG_COMMENT,
-        data: { commentId, songId, playlistId },
-    }
-}
-
 const fetchPlaylists = playlists => {
     // Simulates call to get all playlists for current user
     // We are assuming we have a user signed in
@@ -63,7 +32,7 @@ const fetchPlaylists = playlists => {
 }
 
 // Thunks
-export const AddPlaylistAction = newPlaylist => {
+export const AddPlaylist = newPlaylist => {
     return (dispatch, getState) => {
         const {
             PlaylistsDataReducer: { statePlaylists },
@@ -73,39 +42,20 @@ export const AddPlaylistAction = newPlaylist => {
     }
 }
 
-export const AddPlaylistSong = (songId, playlistId, comment) => {
-    return (dispatch, getState) => {
-        // This is updating PlaylistsDataReducer
-        dispatch(addPlaylistSong(songId, playlistId, comment))
-    }
-}
-
-export const AddSongComment = (commentId, songId, playlistId) => {
-    // We need to pass in playlistId && we need to map this in our playlost state
-    return (dispatch, getState) => {
-        dispatch(addSongComment(commentId, songId, playlistId))
-    }
-}
-
-export const DeletePlaylistAction = playlistId => {
+export const DeletePlaylist = playlistId => {
     return (dispatch, getState) => {
         const {
             PlaylistsDataReducer: { playlists },
         } = getState()
 
-        dispatch(deletePlaylist(playlistId, playlists))
-    }
-}
+        playlists.byId[playlistId].songs.forEach(songId => {
+            // Delete songs from playlist from SongsDataReducer
+            // Delete comments from song from CommentsDataReducer
+            dispatch(DeleteSong(playlistId, songId))
+        })
 
-export const DeletePlaylistSong = (songId, playlistId) => {
-    return (dispatch, getState) => {
-        dispatch(deletePlaylistSong(songId, playlistId))
-    }
-}
-
-export const DeleteSongComment = (commentId, songId, playlistId) => {
-    return (dispatch, getState) => {
-        dispatch(deleteSongComment(commentId, songId, playlistId))
+        // Delete playlist from PlaylistsDataReducer
+        dispatch(deletePlaylist(playlistId))
     }
 }
